requirements.txt
pyserial==3.5
rshell==0.0.31
loguru==0.6.0
paho-mqtt==1.6.1

influxdb-client==1.32.0
